bug fix, new feature microservices

+++ book-user/src/main/java/br/com/biblioteca/bookuser/book/services/DeleteBookServiceImpl.java
@@ -1,9 +1,13 @@
 package br.com.biblioteca.bookuser.book.services;
 
+import br.com.biblioteca.bookuser.book.Book;
 import br.com.biblioteca.bookuser.book.BookRepository;
+import br.com.biblioteca.bookuser.exceptions.BookIntegrityException;
 import br.com.biblioteca.bookuser.exceptions.BookNotDeletedException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
@@ -15,6 +19,10 @@ public class DeleteBookServiceImpl implements DeleteBookService {
     public void delete(Long id) {
         if (!bookRepository.existsById(id)) {
             throw new BookNotDeletedException();
+        }
+        Optional<Book> book = bookRepository.findById(id);
+        if (book.get().getLoanSpecificID() != null) {
+            throw new BookIntegrityException();
         }
         bookRepository.deleteById(id);
     }
