+++ book-user/src/main/java/br/com/biblioteca/bookuser/user/services/DeleteUserAppServiceImpl.java
@@ -1,9 +1,13 @@
 package br.com.biblioteca.bookuser.user.services;
 
+import br.com.biblioteca.bookuser.exceptions.UserAppIntegrityException;
 import br.com.biblioteca.bookuser.exceptions.UserAppNotDeletedException;
+import br.com.biblioteca.bookuser.user.UserApp;
 import br.com.biblioteca.bookuser.user.UserAppRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
@@ -15,6 +19,10 @@ public class DeleteUserAppServiceImpl implements DeleteUserAppService {
     public void delete(Long id) {
         if (!userAppRepository.existsById(id)) {
             throw new UserAppNotDeletedException();
+        }
+        Optional<UserApp> userApp = userAppRepository.findById(id);
+        if (userApp.get().getLoanSpecificID() != null){
+            throw new UserAppIntegrityException();
         }
         userAppRepository.deleteById(id);
     }
